This is radix sort for lists with arbitrary types
This is similar to the stdlib function qsort in that you use some function that can convert a void * type into a long (preferably using positive numbers) and pass that function into rsort so that it has something that is consistent with the rules of radix sort. This is released under the GNU open source license such that you must share any changes you make to the code and publish them with the same license.